Fix NAG Fortran compiler issues across land surface and utility components
- nanMod.F90, clm_varcon.F90, CN_DriverMod.F90: fix NaN/kind handling for NAG
- lsm_routines.F90, StieglitzSnow.F90: fix type/kind issues
- GEOS_IgniGridComp.F90, GEOS_RouteGridComp.F90: NAG compatibility fixes
- makebcs/: fix scalar-vs-array, type mismatch, and OPEN CONVERT issues; add NAG CMake flags
- mk_restarts/: fix NF_DEF_VAR dimids array, kind literals, type mismatches; add NAG CMake flags
- generate_scrip_cube.F90, utils_topo/CMakeLists.txt: NAG fixes for topo utilities
- get_finalID_msk.f90, get_Pfaf_file.f90: fix free-form source and type issues

+++ GEOSagcm_GridComp/GEOSphysics_GridComp/GEOSsurface_GridComp/GEOSroute_GridComp/GEOS_RouteGridComp.F90
@@ -18,6 +18,7 @@ module GEOS_RouteGridCompMod
   use ESMF
   use MAPL_Mod
   use MAPL_ConstantsMod
+  use iso_fortran_env, only: REAL64
   use ROUTING_MODEL,          ONLY: river_routing_hyd
   use reservoirMod,           ONLY: RES_STATE, Reservoir
   use catch_constants,        ONLY: N_pfaf_g => CATCH_N_PFAFS
@@ -545,7 +546,7 @@ contains
       type (ESMF_Grid) :: pfaf_grid
       integer, optional, intent(out) :: rc
       integer :: status
-      real(kind=8), pointer :: centers(:,:)
+      real(REAL64), pointer :: centers(:,:)
       ! create catchment grid and it is tile space
       pfaf_Grid = ESMF_GridCreate(       &
            name='CATCHMENT_GRID',         &
